Refactor location of operands
