+++ src/Passes/Source/CMakeLists.txt
@@ -33,10 +33,12 @@ target_link_libraries(AllocationManager
 
 # Creating the rules library
 file(GLOB RULES_SOURCE RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/Rules/*.cpp)
+file(GLOB RULES_OPRANDS_SOURCE RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ${CMAKE_CURRENT_SOURCE_DIR}/Rules/Oprands/*.cpp)
 
 add_library(Rules 
             SHARED 
-            ${RULES_SOURCE} )
+            ${RULES_SOURCE}
+            ${RULES_OPRANDS_SOURCE} )
 
 target_include_directories(
   Rules
