+++ src/Passes/Source/Rules/OperandPrototype.cpp
@@ -87,89 +87,5 @@ void OperandPrototype::uncapture(Value * /*value*/, Captures &captures) const
   }
 }
 
-CallPattern::CallPattern(String const &name)
-  : name_{name}
-{}
-
-CallPattern::~CallPattern() = default;
-
-bool CallPattern::match(Value *instr, Captures &captures) const
-{
-  auto *call_instr = llvm::dyn_cast<llvm::CallBase>(instr);
-  if (call_instr == nullptr)
-  {
-    return fail(instr, captures);
-  }
-
-  auto target_function = call_instr->getCalledFunction();
-  auto name            = target_function->getName();
-
-  if (name != name_)
-  {
-    return fail(instr, captures);
-  }
-
-  return success(instr, captures);
-}
-
-CallPattern::Child CallPattern::copy() const
-{
-  auto ret = std::make_shared<CallPattern>(name_);
-  ret->copyPropertiesFrom(*this);
-  return std::move(ret);
-}
-
-AnyPattern::AnyPattern()  = default;
-AnyPattern::~AnyPattern() = default;
-bool AnyPattern::match(Value *instr, Captures &captures) const
-{
-  return success(instr, captures);
-}
-
-AnyPattern::Child AnyPattern::copy() const
-{
-  return std::make_shared<AnyPattern>();
-}
-
-template <typename T>
-InstructionPattern<T>::~InstructionPattern() = default;
-template <typename T>
-bool InstructionPattern<T>::match(Value *instr, Captures &captures) const
-{
-  auto *load_instr = llvm::dyn_cast<T>(instr);
-  if (load_instr == nullptr)
-  {
-    return fail(instr, captures);
-  }
-
-  return success(instr, captures);
-}
-
-template <typename T>
-typename InstructionPattern<T>::Child InstructionPattern<T>::copy() const
-{
-  auto ret = std::make_shared<InstructionPattern<T>>();
-  ret->copyPropertiesFrom(*this);
-  return std::move(ret);
-}
-
-// TODO(tfr): This seems to be a bug in LLVM. Template instantiations in
-// a single translation unit is not supposed to reinstantiate across other
-// translation units.
-//
-// However, it is suspecious that htis problem has been around since Clang 8.
-// so this needs more investigation. For now, this work around suffices
-// See
-// https://bugs.llvm.org/show_bug.cgi?id=18733
-// https://stackoverflow.com/questions/56041900/why-does-explicit-template-instantiation-result-in-weak-template-vtables-warning
-// for more information
-#pragma clang diagnostic push
-#pragma clang diagnostic ignored "-Wweak-template-vtables"
-template class InstructionPattern<llvm::StoreInst>;
-template class InstructionPattern<llvm::LoadInst>;
-template class InstructionPattern<llvm::BitCastInst>;
-template class InstructionPattern<llvm::BranchInst>;
-#pragma clang diagnostic pop
-
 }  // namespace quantum
 }  // namespace microsoft
