+++ src/Passes/Source/Rules/ReplacementRule.hpp
@@ -3,7 +3,9 @@
 // Licensed under the MIT License.
 
 #include "Llvm/Llvm.hpp"
-#include "Rules/OperandPrototype.hpp"
+#include "Rules/Operands/Any.hpp"
+#include "Rules/Operands/Call.hpp"
+#include "Rules/Operands/Instruction.hpp"
 
 #include <unordered_map>
 #include <vector>
